+++ tests/examples/models.py
@@ -15,6 +15,7 @@ class SimpleModel(orm.StdModel):
     code = orm.SymbolField(unique = True)
     group = orm.SymbolField(required = False)
     description = orm.CharField()
+    somebytes = orm.ByteField()
     
     objects = CustomManager()